// apps/webapp/src/hooks/morpho/morpho.d.ts
export type MorphoVaultV2Transaction = {
  vault: {
    address: string;
    asset: {
      symbol: string;
      decimals: number;
    };
  };
  type: MorphoTransactionType;
  timestamp: number;
  txHash: string;
  data: {
    assets: string;
  };
};